requirements.txt
numpy==2.1.1
opencv-python==4.11.0.86
opencv-python-headless==4.10.0.84
supervision==0.25.1
ultralytics==8.3.86
pandas==2.2.3
pillow==11.1.0
protobuf==5.29.3
streamlit==1.43.1
